Big focus on GIF generation.
I removed the Python implementation (too slow) and the gnuplot script
(not easy to modify on-the-fly). Configuration is done through
parameters.h, which passes relevant plotting information to gnuplot
through an command executed by the C program (variable "gnuplotCMD"). In
the future, I may add the ability to change rotation speed of the GIF
and support more types of changes to f_c ("cVal") than just a rotating
phasor. jet.pal provides the color scheme for the heatmap.